transfert des fonctions dans les utils pour les test
