+++ shot-n-go_app/src/utils/shoppinglistUtils.mjs
@@ -1,0 +1,18 @@
+export const handleChangeMachine = (
+   machineId,
+   cart,
+   setCart,
+   setSelectedMachineId,
+   fetchMachineShots
+) => {
+   if (
+      cart.length > 0 &&
+      !window.confirm(`Voulez-vous vraiment supprimer votre panier actuel ?`)
+   ) {
+      return;
+   }
+
+   setCart([]);
+   setSelectedMachineId(machineId);
+   fetchMachineShots(machineId);
+};
