+++ shot-n-go_app/src/utils/shotitemUtils.mjs
@@ -1,0 +1,9 @@
+export const lessOneAddAmout = (addAmount, setAddAmount) => {
+  if (addAmount > 0) {
+    setAddAmount(addAmount - 1);
+  }
+};
+
+export const plusOneAddAmout = (addAmount, setAddAmount) => {
+  setAddAmount(addAmount + 1);
+};
